Membresias return data y count

+++ src/Database/Infrastructure/Drizzle/schemas/MembresiaSchema.ts
@@ -4,7 +4,7 @@ export const MembresiaSchema = mysqlTable('membresias', {
   id: int().autoincrement().primaryKey(),
   nombre: varchar({ length: 50 }).notNull(),
   precio: double().notNull(),
-  duracion_meses: int().notNull(),
+  duracion_meses: double().notNull(),
   descripcion: varchar({ length: 200 }).notNull(),
   tipo: varchar({ length: 20 }).notNull(),
 });
